All C unit tests ported and passing in C#

+++ CSharp/src/LiteCore/Interop/C4Document.cs
@@ -36,7 +36,7 @@ namespace LiteCore.Interop
     }
 
     [Flags]
-    public enum C4RevisionFlags : uint
+    public enum C4RevisionFlags : byte
     {
         Deleted = 0x01,
         Leaf = 0x02,
